backend/server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
vaderSentiment>=3.3.2
PyJWT>=2.8.0
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
bcrypt>=4.1.0
cryptography>=42.0.0
